+++ dify_open_webui_adapter.py
@@ -5,32 +5,6 @@ Integrate Open WebUI and Dify by exposing a Dify App
 Supported Open WebUI Version:   v0.7.1
 Supported Dify Version:         1.11.2
 
-User must configure these 2 constant in Python script before use:
-
-``DIFY_BACKEND_API_BASE_URL``: base URL to access Dify Backend Service API
-
-``APP_MODEL_CONFIGS``: a ``list`` of model/app config (each as ``dict``)
-
-example for a single model/app::
-
-    {
-        "key": "...",             # Backend Service API secret key of Dify App
-        "model_id": "model_id1",  # model id as used in Open WebUI
-        "name": "First Model",    # model Name as appeared in Open WebUI
-    }
-
-example for ``APP_MODEL_CONFIGS``::
-
-    APP_MODEL_CONFIGS = [
-        {
-            "key": "...",
-            "model_id": "model_id1",
-        },
-        {
-            ~  # config for 2nd app/model
-        },
-    ]
-
 Q.v. ``https://github.com/kami-lel/dify-open-webui-adapter``
 """
 
@@ -42,10 +16,7 @@ __author__ = "kamiLeL"
 # config  ######################################################################
 DIFY_BACKEND_API_BASE_URL = "https://api.dify.ai/v1"
 
-# Todo: allow additional_inputs: {}
 APP_MODEL_CONFIGS = []
-
-DEBUG_CONVERSATION_ROUND_DIRECT_RESPONSE = False
 
 # end of config  ###############################################################
 
@@ -58,17 +29,15 @@ from pydantic import BaseModel
 import requests
 
 # constants  ===================================================================
+DEBUG_CONVERSATION_ROUND_DIRECT_RESPONSE = False
 OWU_USER_ROLE = "user"
 REQUEST_TIMEOUT = 30
 STREAM_REQUEST_TIMEOUT = 300
 
 # Dify constants  **************************************************************
 DIFY_USER_ROLE = "user"
-# Todo make both configurable in CONFIG
-# in START Node of Workflow in Dify, add an Input Field named 'input'
-DIFY_DEFAULT_INPUT_FIELD_QUERY_KEY = "query"
-# in END Node of Workflow in Dify, add a Output Variable named 'output'
-DIFY_DEFAULT_OUTPUT_VARIABLE_KEY = "output"
+DEFAULT_QUERY_INPUT_FIELD_IDENTIFIER = "query"
+DEFAULT_REPLY_OUTPUT_VARIABLE_IDENTIFIER = "response"
 
 
 # helper Enum  =================================================================
@@ -123,7 +92,7 @@ class OWUModel:
 
         # create app
         if app_type == DifyAppType.WORKFLOW:
-            self.app = WorkflowDifyApp(self)
+            self.app = WorkflowDifyApp(self, app_model_config)
         else:
             self.app = ChatflowDifyApp(self)
 
@@ -409,6 +378,31 @@ class WorkflowDifyApp(BaseDifyApp):
     representing a Workflow App in Dify
     """
 
+    def __init__(self, model, config):
+        super().__init__(model)
+        # read from config  ----------------------------------------------------
+        self.query_identifier = config.get(
+            "query_input_field_identifier",
+            DEFAULT_QUERY_INPUT_FIELD_IDENTIFIER,
+        )
+        self.reply_identifier = config.get(
+            "reply_output_variable_identifier",
+            DEFAULT_REPLY_OUTPUT_VARIABLE_IDENTIFIER,
+        )
+        # read additional input fields
+        self.input_fields = {
+            k: v
+            for k, v in config.items()
+            if k
+            not in (
+                "key",
+                "model_id",
+                "name",
+                "query_input_field_identifier",
+                "reply_output_variable_identifier",
+            )
+        }
+
     @property
     def endpoint_url(self):
         return "{}/workflows/run".format(self.base_url)
@@ -422,9 +416,7 @@ class WorkflowDifyApp(BaseDifyApp):
         response = response_object.json()
 
         try:
-            return response["data"]["outputs"][
-                DIFY_DEFAULT_OUTPUT_VARIABLE_KEY
-            ]
+            return response["data"]["outputs"][self.reply_identifier]
 
         except KeyError as err:
             # Bug test what happens with mismatched key
@@ -445,7 +437,7 @@ class WorkflowDifyApp(BaseDifyApp):
 
     def _create_post_request_payload(self, newest_msg, enable_stream=False):
         payload_dict = {
-            "inputs": {DIFY_DEFAULT_INPUT_FIELD_QUERY_KEY: newest_msg},
+            "inputs": {self.query_identifier: newest_msg, **self.input_fields},
             "response_mode": "streaming" if enable_stream else "blocking",
             "user": DIFY_USER_ROLE,
         }